+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased] 
+
+# [17.103.0-M13] - 2025-06-18
 ### Added
 - Add sql scripts to allow rollback of individual liquibase scripts
 - Add index `stream_status_src_comp_idx` on `stream_status` table to improve stream metrics query